interactive list part done
