+++ src/tasks.js
@@ -1,0 +1,5 @@
+import { getLocalStorage } from './modules/localStorage.js';
+
+const toDoTasks = getLocalStorage();
+
+export default toDoTasks;
